+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - `RGBInt`, an inline class that stores `RGB` colors in a single Int. Create instances directly, or convert to it with `RGB.toRGBInt()`
+- `LinearRGB`, a linearization of the sRGB color space, more suitable for color manipulation. 
 
 ### Changed